Add yezzey virtual index internal relation
TDB use this relation for read-write logic and support
vacuum/truncate with this relation

Relation contains meta-info about every external chunk of data

+++ include/proxy.h
@@ -43,6 +43,6 @@ EXTERNC SMGRFile yezzey_AORelOpenSegFile(char *nspname, char *relname,
                                          int fileMode, int64 modcount);
 EXTERNC int yezzey_FileWrite(SMGRFile file, char *buffer, int amount);
 EXTERNC int yezzey_FileRead(SMGRFile file, char *buffer, int amount);
-EXTERNC int yezzey_FileTruncate(SMGRFile file, int offset);
+EXTERNC int yezzey_FileTruncate(SMGRFile file, int64 offset);
 
 #endif /* YEZZEY_PROXY_H */
